+++ packages/actionsflow-interface/CHANGELOG.md
@@ -5,7 +5,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## 0.0.2 (2020-08-19)
 
-
 ### Features
 
-* 🎸 add actionsflow-interface ([f7ba6a9](https://github.com/actionsflow/actionsflow/commit/f7ba6a91c8083f3379c70735975020386d1dc86a))
+- 🎸 add actionsflow-interface ([f7ba6a9](https://github.com/actionsflow/actionsflow/commit/f7ba6a91c8083f3379c70735975020386d1dc86a))